% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_missing_column.R
\name{add_missing_col}
\alias{add_missing_col}
\title{Add Missing Column}
\usage{
add_missing_col(d, col, type = "logical")
}
\arguments{
\item{d}{data frame}

\item{col}{column name to be added}

\item{type}{column type ('character','integer' or 'numeric', else defaults to 'logical')}
}
\value{
tibble
}
\description{
Add a missing column to a data frame
}
\examples{
# Add a new character column
tibble::tibble(a=1, b=2) |> add_missing_col(col='c',type='character')

Trying to add a column that already exists returns the existing data frame
tibble::tibble(a=1, b=2) |> add_missing_col(col='a',type='character')

}
