% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_proportion.R
\name{add_prop}
\alias{add_prop}
\title{Add Proportion}
\usage{
add_prop(df, col = n, digits = 1)
}
\arguments{
\item{df}{dataframe}

\item{col}{column to be evaluated, default `n`}

\item{digits}{number of decimal places returned, default 1}
}
\value{
tibble
}
\description{
Add Proportion
}
\examples{
tibble::tibble(x=c(2,4,6,8,10)) |> add_prop(x, digits = 2)

}
