% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baen}
\alias{baen}
\title{Differences in flood stages between two stations}
\format{Vector of 33 values}
\source{
D'Agostino and Stephens (1986) page 549.
}
\usage{
baen
}
\description{
Differences in flood stages for two stations on the Fox River, Wisconsin. For testing for Laplace distribution.
}
\keyword{datasets}

