% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beta_data}
\alias{beta_data}
\title{Beta-distributed data}
\format{Vector of 50 observations}
\source{
Generated by package author
}
\usage{
beta_data
}
\description{
Some randomly-generated Beta(1.2,1) data for testing against normality
}
\keyword{datasets}

