% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blis.b}
\alias{blis.b}
\title{Bliss leghorn data series B}
\format{Vector of 21 values}
\source{
D'Agostino and Stephens (1986) page 546.
}
\usage{
blis.b
}
\description{
Weights of 21 leghorn chicks (grams, measured at 21 days). Series A is data \code{leghorn}.
}
\keyword{datasets}

