% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcstats.R
\name{calc.stat}
\alias{calc.stat}
\title{Calculate an EDF statistic}
\usage{
calc.stat(statfun, x, dist = punif, ...)
}
\arguments{
\item{statfun}{Name of one of the EDF statistics (see Details)}

\item{x}{Vector of data to be tested for fit to the distribution}

\item{dist}{Cumulative distribution function of distribution to be fitted}

\item{...}{parameters to distribution}
}
\description{
Calculate test statistic value (and P-value?) for testing fit to specified distribution.
}
\details{
Allowable test statistics are \code{d} (Kolmogorov D), \code{v} (Kuiper V), \code{w2} (Cramer-von Mises W-squared),
  \code{a2} (Anderson-Darling A-squared), \code{u2} (Watson U-squared). Statistics \code{v} and \code{u2} are invariant to choice
  of origin, so can be used for data on a circle (for example, for testing uniformity of times of day).
}
\examples{
data(leghorn)
# test the transformed data for uniformity (default)
calc.stat(a2,leghorn$z1)
# test original data for normality with mean 200 and SD 35
calc.stat("a2",leghorn$x,pnorm,200,35)
# obtain all statistics
my.stats=c("d","v","w2","a2","u2")
sapply(my.stats,calc.stat,leghorn$x,pnorm,200,35)
}

