% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chen}
\alias{chen}
\title{Lethal dose of cinobufagin}
\format{Vector of 25 values}
\source{
D'Agostino and Stephens (1986) page 547.
}
\usage{
chen
}
\description{
Lethal dose of drug cinobufagin in units of 10 mg/kg, as determined by titration
to cardiac arrest in individual etherized cats
}
\keyword{datasets}

