% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{circle}
\alias{circle}
\title{Data on a circle}
\format{Vector of 4 observations}
\source{
D'Agostino and Stephens (1986) page 107
}
\usage{
circle
}
\description{
TIny sample of values on circle of circumference 1.
Used to illustrate that moving these values around the circle (eg.
by subtracting 0.25 from all of them, which, in a clockwise orientation, will change the origin from
North to East) the statistics V and U-squared will not change, even though the others
change.
}
\keyword{datasets}

