% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{combine.eyedata}
\alias{combine.eyedata}
\title{Combine the results from edf.batch}
\usage{
combine.eyedata(batchdata, fields = c("fixations", "saccades", "blinks",
  "messages"))
}
\arguments{
\item{batchdata}{output from \code{\link{edf.batch}}}

\item{fields}{character array of the particular fields you want to combine (default is c('fixations',
'saccades','blinks','messages')). Warning: combining samples will result in extremely large data frames (millions of rows per subject)}
}
\value{
a list with each of the requested fields. Each one is a data frame with all subjects' data. Subjects will
be identified by the "ID" varaiable.
}
\description{
\code{combine.eyedata} can be used to take the results from \code{\link{edf.batch}} and stack
into large data frames for fixations, saccades, etc.
}
\details{
\code{edf.batch} returns a list where each element has fields \code{fixations}, \code{saccades},
\code{blinks}, and (optionaly) \code{samples}. This function takes that result and creates a single
data frame for all subjects' eye data
}
\examples{
\dontrun{

#first load everything:
allbatch <- edf.batch('~/Desktop/edfs','*.edf')

alldata <- combine.eyedata(allbatch)

all_fixations <- alldata$fixations
all_saccades <- alldata$saccades

}
}
\author{
Jason Hubbard, \email{hubbard3@uoregon.edu}
}

