% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edfstats.R
\name{dmod}
\alias{dmod}
\title{Modified D statistic}
\usage{
dmod(x)
}
\arguments{
\item{x}{vector of data that have been probability-integral-transformed}
}
\value{
test statistic value
}
\description{
Calculate modified D statistic (from transformed data) as per formula given in D'Agostino and Stephens (1986) page 105
}

