% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edf_api.R
\name{edf.all}
\alias{edf.all}
\title{Load all data from EDF file}
\usage{
edf.all(EDFfile, samples = FALSE, eventmask = FALSE)
}
\arguments{
\item{EDFfile}{path to an EDF file}

\item{samples}{logical indicating whether to import samples (default=FALSE)}

\item{eventmask}{logical indicating whether to add an \code{\link{eventmask}} to samples (default=FALSE)}
}
\value{
The output will be a list with 4 named elements (fixations, saccades, blinks, and samples)
each element being a data frame
}
\description{
\code{edf.all} returns all of the most common information from
an SR-Research EyeLink EDF file (fixations, saccades, blinks, and samples)
}
\details{
edf.all is useful for obtaining fixations, saccades, blinks, and (optionally)
samples from an EDF in one shot. If you need only 1 of these (i.e., just fixations)
then use \code{\link{edf.events}}, \code{\link{edf.samples}}, \code{\link{edf.messages}}, or
\code{\link{edf.recordings}}. By default it grabs only event data. Use the \code{samples}
 argument to get sample data as well.
}
\examples{
\dontrun{
output <- edf.all('/path/to/file.edf',samples=TRUE)
output$fixations #data frame
output$saccades #another data frame

}
}
\author{
Jason Hubbard, \email{hubbard3@uoregon.edu}
}

