% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{edf.plot}
\alias{edf.plot}
\title{Create a fixation scatterplot from EDF}
\usage{
edf.plot(EDFfile = NULL, fixdata = NULL, outfile = NULL, outlier.rm = F,
  res = NULL, flip = T, theme = "black", crosshairs = T,
  plot.title = "")
}
\arguments{
\item{EDFfile}{path to an EDF file. If you have already-loaded data, leave this as NULL}

\item{fixdata}{optionally you can provide a data frame of already imported fixation data
from \code{\link{edf.events}}. It expects the fields \code{gavx} and \code{gavy}}

\item{outfile}{if provided, it will save the plot as a PDF with this file  name}

\item{outlier.rm}{logical specifying whether to remove outlier fixations (>3SDs from mean)}

\item{res}{array specifying resolution of monitor. Only this region will be plotted.}

\item{flip}{logical indicating whether to flip the y axis so origin is in upper-left (defualt=TRUE).}

\item{theme}{either "black" or "white" specifying the plot theme (yellow on black or black on white)}

\item{crosshairs}{logical whether to plot crosshairs over the median x and y coordinates (dotted red lines)}

\item{plot.title}{title to put at the top of the plot}
}
\description{
This creates a scatterplot of fixations read from an EDF file.
}
\examples{
\dontrun{

edf.plot('/path/to/edffile.edf',outfile='~/Desktop/edfplot.pdf',res=c(1024,768))

#just plot, don't save the PDF
edf.plot('/path/to/edffile.edf',res=c(1024,768),theme='white',plot.title='Fixations')

}
}
\author{
Jason Hubbard, \email{hubbard3@uoregon.edu}
}

