% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{edf.trialcount}
\alias{edf.trialcount}
\title{Determine trial number from an EDF file}
\usage{
edf.trialcount(EDFfile)

edf.trialcount(EDFfile)
}
\arguments{
\item{EDFfile}{path to an EDF file}

\item{EDFfile}{path to an EDF file}
}
\value{
The number of trials in the EDF

The number of trials in the EDF
}
\description{
\code{edf.trialcount} simply returns the number of trials in a given EDF file.

\code{edf.trialcount} simply returns the number of trials in a given EDF file.
}
\details{
\code{edf.trialcount} may be useful if one wants to quickly get the number of trials
from an EDF with little overhead. Trials are defined by the starting and stopping of
recording from the eyetracker. To obtain trial-by-trial data from the EDF, use
\code{\link{edf.trials}}.

\code{edf.trialcount} may be useful if one wants to quickly get the number of trials
from an EDF with little overhead. Trials are defined by the starting and stopping of
recording from the eyetracker. To obtain trial-by-trial data from the EDF, use
\code{\link{edf.trials}}.
}
\examples{
\dontrun{
output <- edf.trialcount('/path/to/file.edf')
output
500
}
\dontrun{
output <- edf.trialcount('/path/to/file.edf')
output
500
}
}
\author{
Jason Hubbard, \email{hubbard3@uoregon.edu}

Jason Hubbard, \email{hubbard3@uoregon.edu}
}

