% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edf_api.R
\name{edf.trials}
\alias{edf.trials}
\title{Load trial-based data from an EDF file}
\usage{
edf.trials(EDFfile, samples = FALSE, eventmask = FALSE,
  event.fields = c("time", "type", "read", "eye", "sttime", "entime", "gstx",
  "gsty", "genx", "geny", "gavx", "gavy", "avel", "pvel"),
  sample.fields = c("time", "gxL", "gyL", "paL", "gxR", "gyR", "paR"))
}
\arguments{
\item{EDFfile}{path to an EDF file}

\item{samples}{logical indicating whether to import samples (default=FALSE)}

\item{eventmask}{logical indicating whether to add an \code{\link{eventmask}} to samples (default=FALSE)}
}
\value{
The output will be a list with 5 named elements (headers,fixations, saccades, blinks, and samples)
each element being a data frame. Headers indicate the starting and stopping point for each trial. Each data
frame will have the variable \code{eyetrial} that indicates the trial number for that event/sample.
}
\description{
\code{edf.trials} loads information from an EDF file in a
trial-by-trial manner. A "trial" is determined by starting
and stopping recording from the eye tracker during an experiment
}
\details{
\code{edf.trials} will load fixations, saccades, blinks, and (optionally)
samples from an EDF similarly to \code{\link{edf.all}}. The difference is that
data are loaded based on separate trials within the recording session. A trial is
determined by stopping and starting recording. This is most useful for merging with
behavioral data recorded elsewhere. Each data frame has an \code{eyetrial} variable
that corresponds to the recording number.
}
\examples{
\dontrun{
output <- edf.trials('/path/to/file.edf',samples=TRUE)
output$fixations #data frame
output$saccades #another data frame

}
}
\author{
Jason Hubbard, \email{hubbard3@uoregon.edu}
}

