% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emea}
\alias{emea}
\title{Heights of maize plants}
\format{Vector of 530 values}
\source{
D'Agostino and Stephens (1986) page 548.
}
\usage{
emea
}
\description{
Heights of maize plants in decimeters (rounded to nearest decimeter)
}
\keyword{datasets}

