% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{epoch.samples}
\alias{epoch.samples}
\title{Epoch sample data}
\usage{
epoch.samples(timelock, samples, sample.field = "paL", epoch = c(-100, 100),
  messages = NULL, eyetrial = FALSE)
}
\arguments{
\item{timelock}{either a string or pattern that matches a messsage sent to Eyelink (e.g., 'STIMONSET', 'TRIAL [0-9]*'),
or a vector of times for events of interest.}

\item{samples}{a data frame of sample data obtained from \code{\link{edf.samples}}, \code{\link{edf.all}}, or
\code{\link{edf.trials}}.}

\item{sample.field}{a character corresponding to the sample field that you are epoching (e.g., 'paL' or 'paR')}

\item{epoch}{the starting and ending samples of the epoch, relative to the event (decault = c(-100, 100))}

\item{messages}{a data frame obtained from \code{\link{edf.messages}}, \code{\link{edf.all}}, or \code{\link{edf.trials}}
which contains the messages sent to Eyelink. Only necessary if \code{timelock} is a character}

\item{eyetrial}{logical indicating whether to include the eyetrial resulting from \code{\link{edf.trials}}. This will be
included as part of the messages data frame if you use \code{\link{edf.trials}}. Default = FALSE}
}
\value{
a list with fields:
\itemize{
\item \code{timelock} a numeric vector of the time-locking events
\item \code{epoch.window} a vector indicating the epoch times you specified with \code{epoch}
\item \code{sample.field} a character indicating the sample field you specified with \code{sample.field}
\item \code{message} if a message is provided, a character vector indicating the message for each epoch
\item \code{eyetrial} if eyetrial=TRUE, then the corresponding eyetrial for each epoch
\item \code{epochs} a matrix of the actual epoched data (events x epoch_width)
}
}
\description{
\code{epoch.samples} is used for time-locking sample data to a particular event and creating
separate epochs.
}
\details{
This is a basic function for epoching sample data (such as pupil diameter) relative to some time-locking event.
The event is indicated either as a message sent to the Eyelink during the experiment, or as a vector of times
of events (which must correspond to \code{time} in the sample data frame).
}
\examples{
\dontrun{

#the easiest way is to use edf.trials:
trials <- edf.trials('/path/to/EDFfile.edf',samples=T,eventmask=T)
trials$samples$paR[trials$samples$blink==1] <- NA #quick blink removal, not perfect
epochdata <- epoch.samples('STIMONSET',trials$samples,sample.field='paR',
                           epoch=c(-500,500),messages=trials$messages,eyetrial=T)
 #plot it!
 matplot(t(epochdata$epochs[1:20,]),type='l')

#if you want to timelock to some type of event:
events <- edf.events('/path/to/EDFfile.edf',type='STARTSACC',fields='sttime')
samples <- edf.samples('/path/to/EDFfile.edf')
epochdata <- epoch.samples(events$sttime,samples,sample.field='paR',epoch=c(-500,1))


}
}
\author{
Jason Hubbard, \email{hubbard3@uoregon.edu}
}

