% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{eventmask}
\alias{eventmask}
\title{Add an event mask to sample data}
\usage{
eventmask(EDFfile, samples = NULL)
}
\arguments{
\item{EDFfile}{path to an EDF file}

\item{samples}{a data frame of sample data, resulting from \code{\link{edf.samples}}}
}
\value{
A data frame of the same sample data, with the variables \code{fixation}, \code{saccade}, and \code{blink} added
to the end.
}
\description{
\code{eventmask} is an internal function that takes sample data from
an EDF file and adds binary variables indicating whether blinks, fixations,
or saccades occurred for each trial
}
\details{
eventmask is called internally from \code{\link{edf.samples}}, \code{\link{edf.trials}}, and \code{\link{edf.all}}
to add 3 binary variables to sample data: \code{fixation}, \code{saccade}, and \code{blink}. Each variable is coded
as 0 or 1 indicating the absence or presence of the fixation/saccade/blink for that sample. This is useful for quickly
indexing sample data based on events (for instance, removing pupil data during blinks). Eyelink (almost) always
records a saccade immediately before and after blinks, and those saccades are included as a part of the blink events.
}
\examples{
\dontrun{

#This is called internally, but if you want to use it manually:
samples <- edf.samples('/path/to/file.edf')

samples_with_mask <- eventmask('/path/to/file.edf',samples)

#usually, you will use it in one of these ways:
trials <- edf.trials('/path/to/file.edf',samples=T,eventmask=T)
all <- edf.all('/path/to/file.edf',samples=T,eventmask=T)
samps <- edf.samples('/path/to/file.edf',eventmask=T)

}
}
\author{
Jason Hubbard, \email{hubbard3@uoregon.edu}
}

