% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_column_names.R
\name{fix_col_names}
\alias{fix_col_names}
\title{Fix Column Names}
\usage{
fix_col_names(d, lower_case = TRUE)
}
\arguments{
\item{d}{dataframe}

\item{lower_case}{Convert to lower case (dafault TRUE)}
}
\value{
dataframe
}
\description{
Fix Column Names
}
