% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_comms.R
\name{fix_comms}
\alias{fix_comms}
\title{Fix Comms}
\usage{
fix_comms(df, col = comms)
}
\arguments{
\item{df}{dataframe}

\item{col}{column to be fixed, default = `comms`}
}
\value{
tibble
}
\description{
Fix Comms
}
\examples{
tibble::tibble(x = c('77123','447456','003451','u192.168.1.1')) |> fix_comms(col=x)

}
