% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mop_portfolio.R
\name{get_mop_portfolio}
\alias{get_mop_portfolio}
\title{Get MOP Portfolio}
\usage{
get_mop_portfolio(supp = FALSE)
}
\arguments{
\item{supp}{Include supplementary fields, Default is FALSE}
}
\value{
tibble
}
\description{
Query the live MOP portfolio
}
