% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leghorn}
\alias{leghorn}
\title{Leghorn data}
\format{A data frame with three variables:
\describe{
  \item{x}{Weight, in grams}
  \item{z1}{Transformed values based on normal with mean 200 and SD 35}
  \item{z2}{Transformed values based on normal using sample mean and SD}
}}
\source{
D'Agostino and Stephens (1986) page 546
}
\usage{
leghorn
}
\description{
Weights of 20 Leghorn chicks. Series A.
}
\keyword{datasets}

