% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_legend.R
\name{move_legend}
\alias{move_legend}
\title{Move Legend}
\usage{
move_legend(position = "none")
}
\arguments{
\item{position}{the position of legends ("none", "left", "right", "bottom", "top", or two-element numeric vector) defaults to "none"}
}
\description{
Move the ggplot2 legend
}
