% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mts_latest_success.R
\name{mts_latest_success}
\alias{mts_latest_success}
\title{Latest MTS Success per Meter}
\usage{
mts_latest_success(
  mts_location = "C:/Users/hale05p/OneDrive - EDF/Projects/MTS/data/"
)
}
\arguments{
\item{mts_location}{Folder location of MTS MTS_success_testid.rds file}
}
\value{
tibble
}
\description{
Latest MTS Success per Meter
}
