% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcstats.R
\name{p.val}
\alias{p.val}
\title{EDF test statistic and P-value}
\usage{
p.val(statfun, x, nsim = 10000, sim = runif, calc = punif, ...)
}
\arguments{
\item{statfun}{Character string naming test to do}

\item{x}{Vector of data}

\item{nsim}{Number of  simulations to run for P-value (default 10,000)}

\item{sim}{Distribution to simulate from for P-value (r-function)}

\item{calc}{Distribution to calculate test statistic for (p-function)}

\item{...}{Parameters for distribution

#' @export}
}
\description{
Calculate EDF test statistic and P-value by simulation
}

