% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvaltab.R
\name{p.val.tab.norm3}
\alias{p.val.tab.norm3}
\title{calculate test statistics and P-values from table 4.7 (normal, case 3)}
\usage{
p.val.tab.norm3(x)
}
\arguments{
\item{x}{vector of data}
}
\value{
data frame of test statistics and P-values
}
\description{
calculate test statistics and P-values from table 4.7 (normal, case 3)
}

