% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvaltab.R
\name{p.val.tab}
\alias{p.val.tab}
\title{Test statistics and P-values from table}
\usage{
p.val.tab(x, calc, ...)
}
\arguments{
\item{x}{vector of data}

\item{calc}{CDF function of distribution to be tested}

\item{...}{parameters for distribution}
}
\value{
data frame of test statistics and P-values
}
\description{
Calculate test statistic and obtain P-value from table 4.2 of D'Agostino and Stephens (1986).
}

