% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_sql.R
\name{pasteSQL}
\alias{pasteSQL}
\title{pasteSQL}
\usage{
pasteSQL(vec)
}
\arguments{
\item{vec}{}
}
\value{
DBI::SQL vector
}
\description{
Collapse a vector into a DBI::SQL character vector (comma separated)
}
\examples{
pasteSQL(letters[c(1:5)])

}
