% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_all.R
\name{print_all}
\alias{print_all}
\title{Print All}
\usage{
print_all(df, n = NULL)
}
\arguments{
\item{df}{dataframe}

\item{n}{Rows to print, default is all rows}
}
\value{
tibble
}
\description{
Print All
}
\examples{
tibble::tibble(x=1:30) |> print_all()
}
