% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvaltab.R
\name{pval.a2}
\alias{pval.a2}
\title{P-value for A-squared from table 4.2}
\usage{
pval.a2(stat)
}
\arguments{
\item{stat}{Test statistic value}
}
\value{
P-value interval (as level of factor)
}
\description{
P-value for A-squared from table 4.2
}
\examples{
pval.a2(1)
pval.a2(2)
pval.a2(3)
pval.a2(5)
pval.a2(10)
}

