% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvaltab.R
\name{pval.d.norm3}
\alias{pval.d.norm3}
\title{P-value for D (unmodified) for normal case 3}
\usage{
pval.d.norm3(stat, n)
}
\arguments{
\item{stat}{Test statistic value}

\item{n}{sample size}
}
\value{
P-value interval (as text)
}
\description{
P-value for D for normality when mean and SD estimated
}

