% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvaltab.R
\name{pval.d}
\alias{pval.d}
\title{P-value for D (modified) from table 4.2}
\usage{
pval.d(stat)
}
\arguments{
\item{stat}{Test statistic value}
}
\value{
P-value interval (as character representation of interval)
}
\description{
P-value for D (modified) from table 4.2
}
\examples{
pval.d(1)
pval.d(1.25)
pval.d(1.5)
pval.d(1.75)
pval.d(2)
}

