% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvaltab.R
\name{pval.tab}
\alias{pval.tab}
\title{Case 0 P-value (interval) from test statistic, list of critical values and list of corresponding P-values.
Helper function, with specific for each test statistic.}
\usage{
pval.tab(stat, cv, pv)
}
\arguments{
\item{stat}{Test statistic value}

\item{cv}{Vector of critical values to look test statistic up in}

\item{pv}{Vector of P-values corresponding to the critical values}
}
\value{
P-value interval (as a level of a factor)
}
\description{
Case 0 P-value (interval) from test statistic, list of critical values and list of corresponding P-values.
Helper function, with specific for each test statistic.
}

