% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvaltab.R
\name{pval.u2}
\alias{pval.u2}
\title{P-value for U-squared (modified) from table 4.2}
\usage{
pval.u2(stat)
}
\arguments{
\item{stat}{Test statistic value}
}
\value{
P-value interval (as level of factor)
}
\description{
P-value for U-squared (modified) from table 4.2
}
\examples{
pval.u2(0.1)
pval.u2(0.175)
pval.u2(0.25)
pval.u2(0.325)
pval.u2(0.4)
}

