% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvaltab.R
\name{pval.v}
\alias{pval.v}
\title{P-value for V (modified) from table 4.2}
\usage{
pval.v(stat)
}
\arguments{
\item{stat}{Test statistic value}
}
\value{
P-value interval (as character representation of interval)
}
\description{
P-value for V (modified) from table 4.2
}
\examples{
pval.v(1.4)
pval.v(1.65)
pval.v(1.9)
pval.v(2.15)
pval.v(2.4)
}

