% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvaltab.R
\name{pval.w2}
\alias{pval.w2}
\title{P-value for W-squared (modified) from table 4.2}
\usage{
pval.w2(stat)
}
\arguments{
\item{stat}{Test statistic value}
}
\value{
P-value interval (as level of factor)
}
\description{
P-value for W-squared (modified) from table 4.2
}
\examples{
pval.w2(0.2)
pval.w2(0.45)
pval.w2(0.7)
pval.w2(0.95)
pval.w2(1.2)
}

