% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_afms.R
\name{query_afms}
\alias{query_afms}
\title{Query AFMS}
\usage{
query_afms(
  q,
  dsn = "AFMS",
  bigint = "integer",
  params = list(),
  show.query = FALSE
)
}
\arguments{
\item{q}{query}

\item{dsn}{driver name, default 'AFMS'}

\item{bigint}{how do we treat bigints, default 'integer'}

\item{params}{optional list of parameters to be passed to the query}

\item{show.query}{show the query being sent, default = FALSE}
}
\value{
tibble
}
\description{
Query AFMS
}
