% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_c2m.R
\name{query_c2m}
\alias{query_c2m}
\title{Query C2M}
\usage{
query_c2m(
  q,
  dsn = "C2M",
  database = "CISADM",
  bigint = "integer",
  LogLevel = 0,
  params = list(),
  show.query = FALSE
)
}
\arguments{
\item{q}{query}

\item{dsn}{driver name, default 'C2M'}

\item{database}{database, default 'CISADM'}

\item{bigint}{how do we treat bigints, defualt 'integer'}

\item{LogLevel}{set to zero to avois printing connection messages}

\item{params}{optional list of parameters to be passed to the query}

\item{show.query}{show the query being sent, default = FALSE}
}
\value{
tibble
}
\description{
Query C2M
}
