% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_x_labs.R
\name{rotate_x_labs}
\alias{rotate_x_labs}
\title{Rotate x labs}
\usage{
rotate_x_labs(angle = 90, vjust = 0.5, hjust = 1, ...)
}
\arguments{
\item{angle}{Angle in \[0, 360\] (default = 90)}

\item{vjust}{Vertical justification in (0,1)}

\item{hjust}{Horizontal justification in (0,1)}

\item{...}{Further named arguments for theme element \code{axis.text.x} passed to \code{ggplot2::element_text()}}
}
\description{
Rotate x labels to be vertical, centered and justified to the x axis
}
