% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcstats.R
\name{sim.stat}
\alias{sim.stat}
\title{Simulate EDF statistics under Case 0}
\usage{
sim.stat(n, statfun, nsim = 10000, dist, ...)
}
\arguments{
\item{n}{Sample size for each simulated test statistic}

\item{statfun}{Name of statistic to be simulated (as in help for \code{calc.stat})}

\item{nsim}{Number of simulations to run (default 10000)}

\item{dist}{Name of R function to draw random samples from distribution (eg. \code{rnorm})}

\item{...}{parameters for distribution}
}
\description{
Obtain simulations of test statistic under assumption that parameters known (for getting P-values).
Case 3 (all parameters unknown) coming up later. Add examples.
}

