% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcstats.R
\name{sim.stat3.1}
\alias{sim.stat3.1}
\title{One simulation of a test statistic from given distribution, parameters to be estimated (case 3)}
\usage{
sim.stat3.1(statfun, n, sim_dist = runif, calc_dist = punif, estim, ...)
}
\arguments{
\item{statfun}{Name of statistic to be simulated (as in help for \code{calc.stat})}

\item{n}{Sample size for each simulated test statistic}

\item{sim_dist}{Name of R function to draw random samples from distribution (eg. \code{rnorm}). Defaults to uniform (\code{runif}).}

\item{calc_dist}{Name of R cumulative distribution to calculate test statistic for (eg. \code{pnorm}). Defaults to uniform (\code{punif}).}

\item{estim}{function parameters for distribution, apparently same for simulation and calculation}

\item{...}{parameter values from data}
}
\description{
One simulation of a test statistic from given distribution, parameters to be estimated (case 3)
}

