% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcstats.R
\name{sim.stat3}
\alias{sim.stat3}
\title{Simulate test statistics under case 3}
\usage{
sim.stat3(statfun, n, nsim = 10000, sim_dist = runif, calc_dist = punif,
  estim, ...)
}
\arguments{
\item{calc_dist}{Name of R function to do probability integral transform with (eg. \code{pnorm}). Defaults to uniform (\code{punif}).}

\item{estim}{Function to calculate parameter estimates from data}

\item{...}{parameters for distribution}
}
\description{
Function to draw random samples from distribution (eg. \code{rnorm}). Defaults to uniform (\code{runif}).
}

