% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcstats.R
\name{test3}
\alias{test3}
\title{Test statistics and P-values under case 3}
\usage{
test3(x, nsim = 10000, calc = punif, sim = runif, estim)
}
\arguments{
\item{x}{vector of data}

\item{nsim}{number of simulations to obtain P-value (default 10,000)}

\item{calc}{Cumulative distribution function of null-hypothesis distribution}

\item{sim}{Function  to generate random sample of values from null-hypothesis distribution}

\item{estim}{Function to calculate parameter estimates from data}
}
\value{
data frame of all test statistic values, labelled by which statistic each one is
}
\description{
Obtain test statistics and P-values for all statistics under case 0, by simulation
}

