% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp_excel.R
\name{tmp_excel}
\alias{tmp_excel}
\title{Write Temp XLSX File}
\usage{
tmp_excel(df)
}
\arguments{
\item{df}{Dataframe to be opened in Excel}
}
\value{
Excel file
}
\description{
Write data to a temporary .xlsx file and open the file. Useful when you need to share a data set with others.
}
