% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edfstats.R
\name{u2}
\alias{u2}
\title{Watson U-squared statistic}
\usage{
u2(x)
}
\arguments{
\item{x}{vector of data that have been probability-integral-transformed}
}
\value{
test statistic value
}
\description{
Calculate U-squared statistic (from transformed data) as per formula given in D'Agostino and Stephens (1986) page 101
}

