% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edfstats.R
\name{u2mod}
\alias{u2mod}
\title{Modified U-squared statistic}
\usage{
u2mod(x)
}
\arguments{
\item{x}{vector of data that have been probability-integral-transformed}
}
\value{
test statistic value
}
\description{
Calculate modified U-squared statistic (from transformed data) as per formula given in D'Agostino and Stephens (1986) page 105
}

