% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edfstats.R
\name{v}
\alias{v}
\title{Kuiper V statistic}
\usage{
v(x)
}
\arguments{
\item{x}{vector of data that have been probability-integral-transformed}
}
\value{
test statistic value
}
\description{
Calculate V statistic (from transformed data) as per formula given in D'Agostino and Stephens (1986) page 101
}

