% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edfstats.R
\name{w2}
\alias{w2}
\title{Cramer-von Mises W-squared statistic}
\usage{
w2(x)
}
\arguments{
\item{x}{vector of data that have been probability-integral-transformed}
}
\value{
test statistic value
}
\description{
Calculate W-squared statistic (from transformed data) as per formula given in D'Agostino and Stephens (1986) page 101
}

