% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_clipboard.R
\name{write_clipboard}
\alias{write_clipboard}
\title{Write Clipboard}
\usage{
write_clipboard(d, size = 32768, sep = "\\t", row.names = FALSE)
}
\arguments{
\item{size}{size of clipbaord to be used (default 32768)}

\item{sep}{separator, default '\t'}

\item{row.names}{default to FALSE}
}
\description{
Write data to the clipboard
}
