% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvaltab.R
\name{p.val.tab.norm3}
\alias{p.val.tab.norm3}
\title{Test statistics and table P-values for normality, parameters estimated}
\usage{
p.val.tab.norm3(x)
}
\arguments{
\item{x}{vector of data}
}
\value{
data frame of test statistics and P-values
}
\description{
calculate test statistics and P-values from table 4.7 of D'Agostino and Stephens (1986)
(normal, case 3)
}

