% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_snowflake.R
\name{query_sf}
\alias{query_sf}
\title{Query Snowflake}
\usage{
query_sf(
  q,
  dsn = "SnowFlake",
  database = "FLK_DUB_DB_DATALAKE_PRD",
  schema = "DIMENSIONAL_IAC",
  bigint = "integer",
  LogLevel = 0,
  params = list(),
  show.query = FALSE
)
}
\arguments{
\item{q}{query}

\item{dsn}{driver name, default 'SnowFlake'}

\item{database}{database, default 'FLK_DUB_DB_DATALAKE_PRD'}

\item{schema}{schema, default 'DIMENSIONAL_IAC'}

\item{bigint}{how do we treat bigints, default 'integer'}

\item{LogLevel}{set to zero to avoid printing connection messages}

\item{params}{optional list of parameters to be passed to the query}

\item{show.query}{show the query being sent, default = FALSE}
}
\value{
tibble
}
\description{
Query Snowflake
}
\examples{
query_sf("select * from dimservicepoint limit 1", show.query = TRUE)

query_sf("select * from dimservicepoint where sk_servicepointid in (?id_)",
          params = list(id_ = DBI::SQL(paste(951495858:951495862, collapse=','))),
          show.query = TRUE)

query_sf("
with a as (
select value as x, *
from table(flatten(input => to_array([?my_values])))
)
select * from a",
params = list(my_values = DBI::SQL(paste(1:10, collapse=','))),
show.query = TRUE)

}
