% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcstats.R
\name{sim.stat}
\alias{sim.stat}
\title{Simulate EDF statistics under Case 0}
\usage{
sim.stat(statfun, n, nsim = 10000, sim_dist = runif, calc_dist = punif,
  ...)
}
\arguments{
\item{statfun}{Name of statistic to be simulated (as in help for \code{calc.stat})}

\item{n}{Sample size for each simulated test statistic}

\item{nsim}{Number of simulations to run (default 10000)}

\item{sim_dist}{Name of R function to draw random samples from distribution (eg. \code{rnorm}). Defaults to uniform (\code{runif}).}

\item{calc_dist}{Name of R function to do probability integral transform with (eg. \code{pnorm}). Defaults to uniform (\code{punif}).}

\item{...}{parameters for distribution}
}
\description{
Obtain simulations of test statistic under assumption that parameters known (for getting P-values).
Case 3 (all parameters unknown) coming up later. Add examples.
}

