% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcstats.R
\name{test0}
\alias{test0}
\title{Test statistics and P-values under case 0}
\usage{
test0(x, nsim = 10000, calc = punif, sim = runif, ...)
}
\description{
Obtain test statistics and P-values for all statistics under case 0
}

